[RISCV] Misc fixes to optimization of vle8.v vse8.v instructions.
This patch fixes how we calculate how many 8 bytes elements per iteration
we can process when we accelerate the vle8.v vse8.v instructions 8 at a time
by calling once vle64.v and vse64.v.
We now set the parameters that describe the size of the elements as they
would be in a 64 bit load/store. That is the log2_esz parameter.
We don't increment the loop counter 'i' by 8 any more as that is already
scaled by log2_esz.
Other fixes take into account the size of the element in order to
scale correctly the number of elements per iteration 'evl' to a 64 bit
load/store.
The patch also makes the 1-byte and the 8-byte strided loops mutually
exclusive as we are not dealing with the tail elements at the moment
but still we want to execute correctly loads and stores of amounts of
data that don't reach 8 bytes.