Add skeleton of new tcg node to do a scalar register based vector load.
This commit adds changes to the RISCV vector translation to call for
a specific node to perform a register offset based vector load
and by passing the relevant SEW information to the tcg function so
to perform correctly the load.
The aim for now is to reach correctness.
This should be applied next to stores.
Optimization will come later.